+++ src/main/java/com/fivemybab/ittabab/store/command/repository/StoreRepository.java
@@ -1,7 +1,7 @@
 package com.fivemybab.ittabab.store.command.repository;
 
-import com.fivemybab.ittabab.store.command.entity.Store;
+import com.fivemybab.ittabab.store.command.domain.aggregate.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepository extends JpaRepository<Store, Integer> {
+public interface StoreRepository extends JpaRepository<Store, Long> {
 }
